finance_bot_patched_fixed/requirements.txt
aiogram>=3.0.0
matplotlib>=3.0.0
pandas>=1.5.0
python-dotenv>=0.21.0
